requirements.txt
asyncpg==0.30.0
bcrypt==4.2.1
python-dotenv==1.0.1
PyJWT==2.10.1
sanic==24.12.0
SQLAlchemy==2.0.38
ujson==5.10.0
